requirements.txt
# IRCTC Tatkal Automation Bot - Python Dependencies

# Core Flask Framework
Flask==2.3.3
Flask-SocketIO==5.3.6

# Browser Automation
playwright==1.40.0

# Async Support
eventlet==0.33.3
python-socketio==5.9.0

# Encryption and Security
cryptography==41.0.7
bcrypt==4.1.2

# Date and Time Handling
python-dateutil==2.8.2
ntplib==0.4.0

# Environment and Configuration
python-dotenv==1.0.0

# Data Processing
requests==2.31.0
urllib3==2.1.0

# JSON and Data Validation
jsonschema==4.20.0
marshmallow==3.20.2

# Database (SQLite support built-in, but adding ORM)
SQLAlchemy==2.0.23

# Scheduling and Background Tasks
APScheduler==3.10.4
celery==5.3.4

# System Monitoring
psutil==5.9.6

# Logging and Monitoring
structlog==23.2.0

# Image Processing (for captcha handling)
Pillow==10.1.0

# HTTP Client
httpx==0.25.2

# File Operations
pathlib2==2.3.7

# Email Support (Optional)
Flask-Mail==0.9.1

# Form Handling and Validation
WTForms==3.1.1
Flask-WTF==1.2.1

# CORS Support
Flask-CORS==4.0.0

# Session Management
Flask-Session==0.5.0

# Password Hashing
argon2-cffi==23.1.0

# Development and Testing (Optional)
pytest==7.4.3
pytest-flask==1.3.0
pytest-asyncio==0.21.1

# Code Quality (Development)
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Documentation (Development)
sphinx==7.2.6

# Deployment
gunicorn==21.2.0
nginx-python-module==1.24.0

# Cloud and AWS Support (Optional)
boto3==1.34.0
botocore==1.34.0

# Performance Monitoring
prometheus-client==0.19.0

# Additional Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1

# Timezone Support
pytz==2023.3

# JSON Web Tokens (Optional)
PyJWT==2.8.0

# Rate Limiting
Flask-Limiter==3.5.0

# API Documentation
flasgger==0.9.7.1

# Configuration Management
dynaconf==3.2.4

# Background Job Processing
redis==5.0.1
rq==1.15.1

# Database Migration (if using more complex DB)
Flask-Migrate==4.0.5

# Testing Framework
coverage==7.3.2
factory-boy==3.3.0

# Development Server
Werkzeug==2.3.7

# Additional Security
itsdangerous==2.1.2

# Memory Usage Optimization
memory-profiler==0.61.0

# File Upload Handling
Flask-Uploads==0.2.1

# Admin Interface (Optional)
Flask-Admin==1.6.1

# Real-time Features
websocket-client==1.6.4

# Caching
Flask-Caching==2.1.0

# Error Tracking (Optional)
sentry-sdk[flask]==1.38.0

# Health Check Endpoints
py-healthcheck==1.10.1